movies_admin/requirements.txt
asgiref==3.5.2
Django==4.0
django-debug-toolbar==3.5.0
django-split-settings==1.1.0
Faker==13.15.1
flake8==4.0.0
mccabe==0.6.1
psycopg2-binary==2.9
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
